#368: export-client is split into internal/external
