+++ pkg/models/notifyupdate.go
@@ -1,0 +1,7 @@
+package models
+
+type NotifyUpdate struct {
+	Name      string `json:"name"`
+	Operation string `json:"operation"`
+}
+
